pyproject.toml
[project]
name = "neo4jgbdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "neo4j-graphrag[openai]>=1.9.1",
    "openai>=1.101.0",
    "python-dotenv>=1.1.1",
]
